Feature: 초기 API 데이터 로드 유무 속성으로 관리
- 속성 필드 batch.init-load.enabled=true,false 로드 유무 관리
